// model/work-session-sorting.ts
export interface WorkSessionSortingBody {
  workSessionId: string;
  dateStart: string;
  timeStart: string;
  productNumber: string;
  lotNumber: string;
  materialNumber: string;
}

export interface WorkSessionSorting {
  id: string;
  workSessionId: string;
  productNumber: string;
  lotNumber: string;
  materialNumber: string;
  dateStart: string;
  timeStart: string;
  remark: string | null;
  dateComplete: string | null;
  timeComplete: string | null;
  status: number;
}

export interface WorkSessionSortingResponse {
  workSessionSorting: WorkSessionSorting;
}

export interface CompleteWorkSessionSorting {
  id: string;
  dateComplete: string;
  timeComplete: string;
}

export interface WorkSessionSortingByWsId {
  id: string;
  workSessionId: string;
  productNumber: string;
  lotNumber: string;
  materialNumber: string;
  dateStart: string;
  timeStart: string;
  remark: string | null;
  dateComplete: string | null;
  timeComplete: string | null;
  status: number;
}

export interface WorkSessionSortingListResponse {
  workSessionSortings: WorkSessionSortingByWsId[];
}